rename atomic blocks and components to be more generic (drop list)
