+++ assets/js/atomic/blocks/product/index.js
@@ -1,0 +1,7 @@
+export { default as ProductTitle } from './title';
+export { default as ProductPrice } from './price';
+export { default as ProductImage } from './image';
+export { default as ProductRating } from './rating';
+export { default as ProductButton } from './button';
+export { default as ProductSummary } from './summary';
+export { default as ProductSaleBadge } from './sale-badge';
