+++ assets/js/atomic/blocks/product/shared-config.js
