--- assets/js/atomic/components/product-list/index.js
@@ -1,7 +1,0 @@
-export { default as ProductListButton } from './button';
-export { default as ProductListImage } from './image';
-export { default as ProductListRating } from './rating';
-export { default as ProductListTitle } from './title';
-export { default as ProductListPrice } from './price';
-export { default as ProductListSummary } from './summary';
-export { default as ProductListSaleBadge } from './sale-badge';
